Python/Case2/module01/requirements01-2.txt
Flask==0.12.2
azure-iot-device~=2.0.0